requirements.txt
# Core Dependencies
pandas==2.1.1
numpy==1.26.1

# Data Processing
python-dateutil==2.8.2

# Reddit API
praw==7.7.1
prawcore==2.4.0

# AWS SDK
boto3==1.31.64
botocore==1.31.64
s3fs==2023.10.0

# Apache Airflow
apache-airflow==2.7.2
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-http==4.6.0

# Machine Learning & NLP
nltk==3.8.1
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Jupyter Notebooks
jupyter==1.0.0
ipykernel==6.26.0
notebook==7.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.49

# HTTP & Requests
requests==2.31.0
urllib3==1.26.18

# Configuration
configparser==6.0.0

# Logging
colorlog==6.8.0
